refactor: split InterviewsTab and extract shared DayTimeline component
- Extract InterviewCard → components/interviews/InterviewCard.tsx
- Extract InterviewForm + makeEmptyForm → components/interviews/InterviewForm.tsx
- Extract shared VIBE_OPTIONS / FEELING_OPTIONS → components/interviews/interviewDisplayOptions.ts
- InterviewsTab.tsx reduced from ~1050 lines to ~290 lines of orchestration logic
- Extract day-rail/timeline layout into shared DayTimeline component, removing
  duplicated JSX from both InterviewsTab and InterviewsPage
