+++ frontend/src/components/InterviewsPage.tsx
@@ -19,6 +19,7 @@ import { useSnackbar } from "../useSnackbar";
 import { formatTime } from "../jobUtils";
 import MarkdownSnippet from "./MarkdownSnippet";
 import CompanyLogo from "./CompanyLogo";
+import DayTimeline from "./DayTimeline";
 import PageSpinner from "./PageSpinner";
 import type { EnrichedInterview, InterviewVibe } from "../types";
 
@@ -305,7 +306,6 @@ export default function InterviewsPage() {
 				{!loading &&
 					grouped.length > 0 &&
 					grouped.map(({ label, items, isPast }) => {
-						// Group by calendar day
 						const dayMap = new Map<string, EnrichedInterview[]>();
 						for (const iv of items) {
 							const key = new Date(iv.interview_dttm).toDateString();
@@ -334,76 +334,18 @@ export default function InterviewsPage() {
 										No interviews this week
 									</Typography>
 								) : (
-									[...dayMap.entries()].map(([dateStr, dayItems]) => {
-										const d = new Date(dateStr);
-										const isToday =
-											d.toDateString() === new Date().toDateString();
-										const weekday = d.toLocaleDateString("en-US", {
-											weekday: "short",
-										});
-										const dateLabel = d.toLocaleDateString("en-US", {
-											month: "short",
-											day: "numeric",
-										});
-										return (
-											<Box
-												key={dateStr}
-												sx={{ display: "flex", alignItems: "stretch", mb: 1.5 }}
-											>
-												{/* Left rail: day label */}
-												<Box
-													sx={{
-														width: 48,
-														flexShrink: 0,
-														textAlign: "right",
-														pr: 1.5,
-														pt: 0.5,
-													}}
-												>
-													<Typography
-														variant="caption"
-														fontWeight={600}
-														color={isToday ? "primary.main" : "text.secondary"}
-														sx={{ display: "block", lineHeight: 1.3 }}
-													>
-														{weekday}
-													</Typography>
-													<Typography
-														variant="caption"
-														color={isToday ? "primary.main" : "text.disabled"}
-														sx={{
-															display: "block",
-															lineHeight: 1.3,
-															fontSize: "0.65rem",
-														}}
-													>
-														{dateLabel}
-													</Typography>
-												</Box>
-												{/* Vertical line */}
-												<Box
-													sx={{
-														width: "2px",
-														flexShrink: 0,
-														mr: 1.5,
-														borderRadius: "2px",
-														bgcolor: isToday ? "primary.light" : "divider",
-													}}
+									[...dayMap.entries()].map(([dateStr, dayItems]) => (
+										<DayTimeline key={dateStr} dateStr={dateStr}>
+											{dayItems.map((iv) => (
+												<InterviewRow
+													key={iv.id}
+													interview={iv}
+													onJobClick={() => navigate(`/jobs/${iv.job.id}`)}
+													dimmed={isPast}
 												/>
-												{/* Cards */}
-												<Box sx={{ flex: 1, minWidth: 0 }}>
-													{dayItems.map((iv) => (
-														<InterviewRow
-															key={iv.id}
-															interview={iv}
-															onJobClick={() => navigate(`/jobs/${iv.job.id}`)}
-															dimmed={isPast}
-														/>
-													))}
-												</Box>
-											</Box>
-										);
-									})
+											))}
+										</DayTimeline>
+									))
 								)}
 							</Box>
 						);
