+++ frontend/src/components/InterviewsTab.tsx
@@ -1,152 +1,13 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
-import {
-	Box,
-	Button,
-	Chip,
-	CircularProgress,
-	IconButton,
-	MenuItem,
-	TextField,
-	ToggleButton,
-	ToggleButtonGroup,
-	Tooltip,
-	Typography,
-} from "@mui/material";
+import { Box, Button, CircularProgress, Typography } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
-import BusinessIcon from "@mui/icons-material/Business";
-import DeleteIcon from "@mui/icons-material/Delete";
-import EditIcon from "@mui/icons-material/Edit";
-import PhoneIcon from "@mui/icons-material/Phone";
-import QuizOutlinedIcon from "@mui/icons-material/QuizOutlined";
 import { api } from "../api";
-import {
-	INTERVIEW_MAX_LENGTHS,
-	INTERVIEW_STAGE_LABELS,
-	INTERVIEW_TYPE_LABELS,
-} from "../constants";
-import type {
-	Interview,
-	InterviewFeeling,
-	InterviewFormData,
-	InterviewResult,
-	InterviewStage,
-	InterviewType,
-	InterviewVibe,
-	JobStatus,
-} from "../types";
-import MarkdownField from "./MarkdownField";
-import MarkdownSnippet from "./MarkdownSnippet";
+import { INTERVIEW_MAX_LENGTHS } from "../constants";
+import type { Interview, InterviewFormData, JobStatus } from "../types";
 import QuestionSubView from "./QuestionSubView";
-import { formatTime } from "../jobUtils";
-
-const VIBE_OPTIONS: {
-	value: InterviewVibe;
-	emoji: string;
-	label: string;
-	selectedBg: string;
-	selectedColor: string;
-	chipBg: string;
-	chipColor: string;
-}[] = [
-	{
-		value: "casual",
-		emoji: "☕",
-		label: "Casual",
-		selectedBg: "#e3f2fd",
-		selectedColor: "#1565c0",
-		chipBg: "#e3f2fd",
-		chipColor: "#1565c0",
-	},
-	{
-		value: "intense",
-		emoji: "⚡",
-		label: "Intense",
-		selectedBg: "#fff3e0",
-		selectedColor: "#e65100",
-		chipBg: "#fff3e0",
-		chipColor: "#e65100",
-	},
-];
-
-const RESULT_CHIP_SX: Record<InterviewResult, object> = {
-	passed: { bgcolor: "#e8f5e9", color: "#2e7d32" },
-	failed: { bgcolor: "#ffebee", color: "#c62828" },
-};
-
-const RESULT_LABELS: Record<InterviewResult, string> = {
-	passed: "✓ Passed",
-	failed: "✗ Failed",
-};
-
-const FEELING_OPTIONS: {
-	value: InterviewFeeling;
-	emoji: string;
-	label: string;
-	selectedBg: string;
-	selectedColor: string;
-	chipBg: string;
-	chipColor: string;
-}[] = [
-	{
-		value: "aced",
-		emoji: "🌟",
-		label: "Aced",
-		selectedBg: "#e8f5e9",
-		selectedColor: "#1b5e20",
-		chipBg: "#e8f5e9",
-		chipColor: "#1b5e20",
-	},
-	{
-		value: "pretty_good",
-		emoji: "👍",
-		label: "Pretty good",
-		selectedBg: "#f1f8e9",
-		selectedColor: "#33691e",
-		chipBg: "#f1f8e9",
-		chipColor: "#33691e",
-	},
-	{
-		value: "meh",
-		emoji: "😐",
-		label: "Meh",
-		selectedBg: "#fff8e1",
-		selectedColor: "#f57f17",
-		chipBg: "#fff8e1",
-		chipColor: "#f57f17",
-	},
-	{
-		value: "struggled",
-		emoji: "😬",
-		label: "Struggled",
-		selectedBg: "#fff3e0",
-		selectedColor: "#e65100",
-		chipBg: "#fff3e0",
-		chipColor: "#e65100",
-	},
-	{
-		value: "flunked",
-		emoji: "💀",
-		label: "Flunked",
-		selectedBg: "#ffebee",
-		selectedColor: "#b71c1c",
-		chipBg: "#ffebee",
-		chipColor: "#b71c1c",
-	},
-];
-
-function makeEmptyForm(jobStatus: JobStatus): InterviewFormData {
-	const isEarly = jobStatus === "Not started" || jobStatus === "Applied";
-	return {
-		interview_dttm: "",
-		interview_interviewers: null,
-		interview_notes: null,
-		interview_stage: isEarly ? "phone_screen" : "onsite",
-		interview_type: isEarly ? "recruiter_call" : null,
-		interview_vibe: null,
-		interview_result: null,
-		interview_feeling: null,
-	};
-}
+import DayTimeline from "./DayTimeline";
+import InterviewCard from "./interviews/InterviewCard";
+import InterviewForm, { makeEmptyForm } from "./interviews/InterviewForm";
 
 type Mode = "list" | "add" | { editId: number } | { confirmDeleteId: number };
 
@@ -426,7 +287,6 @@ export default function InterviewsTab({
 			return null;
 		}
 
-		// Group by calendar day
 		const dayMap = new Map<string, Interview[]>();
 		for (const iv of sectionInterviews) {
 			const key = new Date(iv.interview_dttm).toDateString();
@@ -445,67 +305,11 @@ export default function InterviewsTab({
 				>
 					{title}
 				</Typography>
-				{[...dayMap.entries()].map(([dateStr, dayInterviews]) => {
-					const d = new Date(dateStr);
-					const isToday = d.toDateString() === new Date().toDateString();
-					const weekday = d.toLocaleDateString("en-US", { weekday: "short" });
-					const dateLabel = d.toLocaleDateString("en-US", {
-						day: "numeric",
-						month: "short",
-					});
-
-					return (
-						<Box
-							key={dateStr}
-							sx={{ alignItems: "stretch", display: "flex", mb: 1.5 }}
-						>
-							{/* Left rail: day label */}
-							<Box
-								sx={{
-									flexShrink: 0,
-									pr: 1.5,
-									pt: 0.5,
-									textAlign: "right",
-									width: 48,
-								}}
-							>
-								<Typography
-									variant="caption"
-									fontWeight={600}
-									color={isToday ? "primary.main" : "text.secondary"}
-									sx={{ display: "block", lineHeight: 1.3 }}
-								>
-									{weekday}
-								</Typography>
-								<Typography
-									variant="caption"
-									color={isToday ? "primary.main" : "text.disabled"}
-									sx={{
-										display: "block",
-										fontSize: "0.65rem",
-										lineHeight: 1.3,
-									}}
-								>
-									{dateLabel}
-								</Typography>
-							</Box>
-							{/* Vertical line */}
-							<Box
-								sx={{
-									bgcolor: isToday ? "primary.light" : "divider",
-									borderRadius: "2px",
-									flexShrink: 0,
-									mr: 1.5,
-									width: "2px",
-								}}
-							/>
-							{/* Cards */}
-							<Box sx={{ flex: 1, minWidth: 0 }}>
-								{dayInterviews.map((iv) => renderItem(iv, questionsDisabled))}
-							</Box>
-						</Box>
-					);
-				})}
+				{[...dayMap.entries()].map(([dateStr, dayInterviews]) => (
+					<DayTimeline key={dateStr} dateStr={dateStr}>
+						{dayInterviews.map((iv) => renderItem(iv, questionsDisabled))}
+					</DayTimeline>
+				))}
 			</Box>
 		);
 	};
@@ -578,471 +382,6 @@ export default function InterviewsTab({
 						<QuestionSubView jobId={jobId} interview={displayedInterview} />
 					)}
 				</Box>
-			</Box>
-		</Box>
-	);
-}
-
-function InterviewCard({
-	interview,
-	questionCount,
-	questionsDisabled,
-	onEdit,
-	onDelete,
-	onViewQuestions,
-}: {
-	interview: Interview;
-	questionCount: number;
-	questionsDisabled: boolean;
-	onEdit: () => void;
-	onDelete: () => void;
-	onViewQuestions: () => void;
-}) {
-	const TypeIcon =
-		interview.interview_stage === "phone_screen" ? PhoneIcon : BusinessIcon;
-	const typeLabel = INTERVIEW_STAGE_LABELS[interview.interview_stage];
-	const isPast = new Date(interview.interview_dttm) < new Date();
-	const feelingOption = interview.interview_feeling
-		? FEELING_OPTIONS.find((o) => o.value === interview.interview_feeling)
-		: null;
-
-	return (
-		<Box
-			sx={{
-				border: "1px solid",
-				borderColor: "divider",
-				borderRadius: 1,
-				mb: 1,
-				overflow: "hidden",
-				p: 1.5,
-			}}
-		>
-			<Box sx={{ alignItems: "flex-start", display: "flex", gap: 1 }}>
-				<TypeIcon
-					sx={{
-						color: "text.secondary",
-						flexShrink: 0,
-						fontSize: 16,
-						mt: 0.25,
-					}}
-				/>
-				<Box sx={{ flex: 1, minWidth: 0 }}>
-					<Box
-						sx={{
-							alignItems: "center",
-							display: "flex",
-							flexWrap: "wrap",
-							gap: 1,
-						}}
-					>
-						<Typography variant="body2" fontWeight={600}>
-							{typeLabel}
-						</Typography>
-						<Typography variant="body2" color="text.secondary">
-							&middot; {formatTime(interview.interview_dttm)}
-						</Typography>
-						{interview.interview_type && (
-							<Chip
-								label={INTERVIEW_TYPE_LABELS[interview.interview_type]}
-								size="small"
-								sx={{ bgcolor: "#f3e5f5", color: "#6a1b9a" }}
-							/>
-						)}
-						{interview.interview_vibe &&
-							(() => {
-								const opt = VIBE_OPTIONS.find(
-									(o) => o.value === interview.interview_vibe,
-								)!;
-								return (
-									<Chip
-										label={`${opt.emoji} ${opt.label}`}
-										size="small"
-										sx={{ bgcolor: opt.chipBg, color: opt.chipColor }}
-									/>
-								);
-							})()}
-						{isPast && interview.interview_result && (
-							<Chip
-								label={RESULT_LABELS[interview.interview_result]}
-								size="small"
-								sx={RESULT_CHIP_SX[interview.interview_result]}
-							/>
-						)}
-						{isPast && feelingOption && (
-							<Chip
-								label={`${feelingOption.emoji} ${feelingOption.label}`}
-								size="small"
-								sx={{
-									bgcolor: feelingOption.chipBg,
-									color: feelingOption.chipColor,
-								}}
-							/>
-						)}
-					</Box>
-					{interview.interview_interviewers && (
-						<Typography
-							variant="caption"
-							color="text.secondary"
-							sx={{ display: "block" }}
-						>
-							{interview.interview_interviewers}
-						</Typography>
-					)}
-					{interview.interview_notes && (
-						<Typography
-							variant="caption"
-							color="text.secondary"
-							sx={{
-								display: "block",
-								mt: 0.25,
-								overflow: "hidden",
-								textOverflow: "ellipsis",
-								whiteSpace: "nowrap",
-							}}
-						>
-							<MarkdownSnippet text={interview.interview_notes} />
-						</Typography>
-					)}
-					<Button
-						size="small"
-						startIcon={<QuizOutlinedIcon sx={{ fontSize: 14 }} />}
-						onClick={onViewQuestions}
-						disabled={questionsDisabled}
-						sx={{ minWidth: 0, mt: 0.5, p: 0, textTransform: "none" }}
-					>
-						{questionCount > 0 ? `Questions (${questionCount})` : "Questions"}
-					</Button>
-				</Box>
-				<Box sx={{ display: "flex", flexShrink: 0 }}>
-					<Tooltip title="Edit interview">
-						<IconButton
-							size="small"
-							onClick={onEdit}
-							aria-label="Edit interview"
-						>
-							<EditIcon fontSize="small" />
-						</IconButton>
-					</Tooltip>
-					<Tooltip title="Delete interview">
-						<IconButton
-							size="small"
-							onClick={onDelete}
-							aria-label="Delete interview"
-						>
-							<DeleteIcon fontSize="small" />
-						</IconButton>
-					</Tooltip>
-				</Box>
-			</Box>
-		</Box>
-	);
-}
-
-interface FormProps {
-	data: InterviewFormData;
-	onChange: <K extends keyof InterviewFormData>(
-		field: K,
-		value: InterviewFormData[K],
-	) => void;
-	onSave: () => void;
-	onCancel: () => void;
-	saving: boolean;
-	error: string | null;
-}
-
-function InterviewForm({
-	data,
-	onChange,
-	onSave,
-	onCancel,
-	saving,
-	error,
-}: FormProps) {
-	const isPast =
-		Boolean(data.interview_dttm) && new Date(data.interview_dttm) < new Date();
-
-	return (
-		<Box
-			sx={{
-				border: "1px solid",
-				borderColor: "primary.light",
-				borderRadius: 1,
-				mb: 1,
-				p: 1.5,
-			}}
-		>
-			<Box sx={{ display: "flex", flexWrap: "wrap", gap: 1.5, mb: 1.5 }}>
-				<TextField
-					select
-					label="Stage"
-					value={data.interview_stage}
-					onChange={(e) =>
-						onChange("interview_stage", e.target.value as InterviewStage)
-					}
-					size="small"
-					sx={{ minWidth: 140 }}
-				>
-					{(
-						Object.entries(INTERVIEW_STAGE_LABELS) as [InterviewStage, string][]
-					).map(([value, label]) => (
-						<MenuItem key={value} value={value}>
-							{label}
-						</MenuItem>
-					))}
-				</TextField>
-				<TextField
-					label="Date & Time"
-					type="datetime-local"
-					value={data.interview_dttm}
-					onChange={(e) => onChange("interview_dttm", e.target.value)}
-					size="small"
-					required
-					slotProps={{ inputLabel: { shrink: true } }}
-					sx={{ minWidth: 220 }}
-				/>
-			</Box>
-			<TextField
-				label="Interviewers"
-				value={data.interview_interviewers ?? ""}
-				onChange={(e) =>
-					onChange("interview_interviewers", e.target.value || null)
-				}
-				size="small"
-				fullWidth
-				placeholder="e.g. Jane Smith, Bob Lee"
-				sx={{ mb: 1.5 }}
-				slotProps={{
-					htmlInput: {
-						maxLength: INTERVIEW_MAX_LENGTHS.interview_interviewers,
-					},
-				}}
-			/>
-			<Box sx={{ mb: 1.5 }}>
-				<TextField
-					select
-					label="Type"
-					value={data.interview_type ?? ""}
-					onChange={(e) =>
-						onChange(
-							"interview_type",
-							(e.target.value || null) as InterviewType | null,
-						)
-					}
-					size="small"
-					sx={{ minWidth: 160 }}
-				>
-					<MenuItem value="">
-						<em>None</em>
-					</MenuItem>
-					{(
-						Object.entries(INTERVIEW_TYPE_LABELS) as [InterviewType, string][]
-					).map(([value, label]) => (
-						<MenuItem key={value} value={value}>
-							{label}
-						</MenuItem>
-					))}
-				</TextField>
-			</Box>
-			<Box sx={{ mb: 1.5 }}>
-				<MarkdownField
-					label="Notes"
-					value={data.interview_notes}
-					onChange={(v) => onChange("interview_notes", v)}
-				/>
-			</Box>
-			<Box sx={{ mb: 1.5 }}>
-				<Typography
-					variant="caption"
-					color="text.secondary"
-					sx={{ display: "block", fontWeight: 600, mb: 0.75 }}
-				>
-					After the interview
-				</Typography>
-				<Box sx={{ display: "flex", flexWrap: "wrap", gap: 3 }}>
-					<Box
-						sx={{
-							alignItems: "center",
-							display: "flex",
-							flexDirection: "column",
-							gap: 0.5,
-						}}
-					>
-						<ToggleButtonGroup
-							exclusive
-							size="small"
-							value={data.interview_vibe}
-							disabled={!isPast}
-							onChange={(_, v) =>
-								onChange("interview_vibe", v as InterviewVibe | null)
-							}
-						>
-							{VIBE_OPTIONS.map((opt) => (
-								<Tooltip key={opt.value} title={opt.label}>
-									<ToggleButton
-										value={opt.value}
-										aria-label={opt.label}
-										sx={{
-											fontSize: "1.3rem",
-											lineHeight: 1,
-											px: 1.25,
-											py: 0.75,
-											"&.Mui-selected": {
-												bgcolor: opt.selectedBg,
-												color: opt.selectedColor,
-												"&:hover": {
-													bgcolor: opt.selectedBg,
-													filter: "brightness(0.95)",
-												},
-											},
-										}}
-									>
-										{opt.emoji}
-									</ToggleButton>
-								</Tooltip>
-							))}
-						</ToggleButtonGroup>
-						<Typography variant="caption" color="text.secondary">
-							Vibe
-						</Typography>
-					</Box>
-					<Box
-						sx={{
-							alignItems: "center",
-							display: "flex",
-							flexDirection: "column",
-							gap: 0.5,
-						}}
-					>
-						<ToggleButtonGroup
-							exclusive
-							size="small"
-							value={data.interview_feeling}
-							disabled={!isPast}
-							onChange={(_, v) =>
-								onChange("interview_feeling", v as InterviewFeeling | null)
-							}
-						>
-							{FEELING_OPTIONS.map((opt) => (
-								<Tooltip key={opt.value} title={opt.label}>
-									<ToggleButton
-										value={opt.value}
-										aria-label={opt.label}
-										sx={{
-											fontSize: "1.3rem",
-											lineHeight: 1,
-											px: 1.25,
-											py: 0.75,
-											"&.Mui-selected": {
-												bgcolor: opt.selectedBg,
-												color: opt.selectedColor,
-												"&:hover": {
-													bgcolor: opt.selectedBg,
-													filter: "brightness(0.95)",
-												},
-											},
-										}}
-									>
-										{opt.emoji}
-									</ToggleButton>
-								</Tooltip>
-							))}
-						</ToggleButtonGroup>
-						<Typography variant="caption" color="text.secondary">
-							Feeling
-						</Typography>
-					</Box>
-					<Box
-						sx={{
-							alignItems: "center",
-							display: "flex",
-							flexDirection: "column",
-							gap: 0.5,
-						}}
-					>
-						<ToggleButtonGroup
-							exclusive
-							size="small"
-							value={data.interview_result}
-							disabled={!isPast}
-							onChange={(_, v) =>
-								onChange("interview_result", v as InterviewResult | null)
-							}
-						>
-							<Tooltip title="Passed">
-								<ToggleButton
-									value="passed"
-									aria-label="Passed"
-									sx={{
-										fontSize: "1.3rem",
-										lineHeight: 1,
-										px: 1.25,
-										py: 0.75,
-										"&.Mui-selected": {
-											bgcolor: "#e8f5e9",
-											color: "#2e7d32",
-											"&:hover": { bgcolor: "#c8e6c9" },
-										},
-									}}
-								>
-									✅
-								</ToggleButton>
-							</Tooltip>
-							<Tooltip title="Failed">
-								<ToggleButton
-									value="failed"
-									aria-label="Failed"
-									sx={{
-										fontSize: "1.3rem",
-										lineHeight: 1,
-										px: 1.25,
-										py: 0.75,
-										"&.Mui-selected": {
-											bgcolor: "#ffebee",
-											color: "#c62828",
-											"&:hover": { bgcolor: "#ffcdd2" },
-										},
-									}}
-								>
-									❌
-								</ToggleButton>
-							</Tooltip>
-						</ToggleButtonGroup>
-						<Typography variant="caption" color="text.secondary">
-							Result
-						</Typography>
-					</Box>
-				</Box>
-				{!isPast && data.interview_dttm && (
-					<Typography
-						variant="caption"
-						color="text.disabled"
-						sx={{ display: "block", mt: 0.5 }}
-					>
-						Available once the interview date has passed
-					</Typography>
-				)}
-			</Box>
-			{error && (
-				<Typography
-					variant="caption"
-					color="error"
-					sx={{ display: "block", mb: 1 }}
-				>
-					{error}
-				</Typography>
-			)}
-			<Box sx={{ display: "flex", gap: 1, justifyContent: "flex-end" }}>
-				<Button size="small" onClick={onCancel} disabled={saving}>
-					Cancel
-				</Button>
-				<Button
-					size="small"
-					variant="contained"
-					onClick={onSave}
-					disabled={saving}
-				>
-					Save Interview
-				</Button>
 			</Box>
 		</Box>
 	);
